[Fix] 시간 설정 윤년으로 변경

+++ Assets/01.Scripts/97.Controller/TimerController.cs
@@ -14,7 +14,7 @@ public class TimerController : MonoBehaviour
         timerUI.SetDay(timerModel.GetFormatDay());
     }
 
-    public void DayChange(float day)
+    public void DayChange(int day)
     {
         timerModel.DayChange(day);
         ChangeDayUI();
